Moved all command stuff to a new package
