+++ src/main/bot-kt/org/kamiblue/botkt/command/commands/DiscussCommand.kt
@@ -1,8 +1,9 @@
-package org.kamiblue.botkt.commands
+package org.kamiblue.botkt.command.commands
 
 import net.ayataka.kordis.entity.server.permission.PermissionSet
 import net.ayataka.kordis.entity.server.permission.overwrite.RolePermissionOverwrite
-import org.kamiblue.botkt.*
+import org.kamiblue.botkt.PermissionTypes
+import org.kamiblue.botkt.command.*
 import org.kamiblue.botkt.utils.Colors
 
 // TODO: make this not hardcoded
