+++ src/main/bot-kt/org/kamiblue/botkt/command/commands/ExampleCommand.kt
@@ -1,6 +1,6 @@
-package org.kamiblue.botkt.commands
+package org.kamiblue.botkt.command.commands
 
-import org.kamiblue.botkt.*
+import org.kamiblue.botkt.command.*
 
 /**
  * @author l1ving
