+++ src/main/bot-kt/org/kamiblue/botkt/command/commands/IssueCommand.kt
@@ -1,4 +1,4 @@
-package org.kamiblue.botkt.commands
+package org.kamiblue.botkt.command.commands
 
 import kotlinx.coroutines.delay
 import net.ayataka.kordis.entity.message.Message
@@ -9,6 +9,7 @@ import net.ayataka.kordis.event.events.message.MessageReceiveEvent
 import net.ayataka.kordis.event.events.message.ReactionAddEvent
 import org.kamiblue.botkt.*
 import org.kamiblue.botkt.Permissions.hasPermission
+import org.kamiblue.botkt.command.*
 import org.kamiblue.botkt.utils.Colors
 import org.kamiblue.botkt.utils.GitHubUtils
 import org.kamiblue.botkt.utils.MessageSendUtils.error
@@ -73,7 +74,8 @@ object IssueCommand : Command("issue") {
                         val title = split.getOrNull(0)
                         val body = split.getOrNull(1)
 
-                        val formattedIssue = "Created by: ${message.author?.name?.toHumanReadable()} (${message.author?.mention})\n\n$body"
+                        val formattedIssue =
+                            "Created by: ${message.author?.name?.toHumanReadable()} (${message.author?.mention})\n\n$body"
                         val issue = Issue(title = title, body = formattedIssue)
 
                         val issueChannel = ConfigManager.readConfig<UserConfig>(ConfigType.USER, false)
@@ -86,7 +88,13 @@ object IssueCommand : Command("issue") {
 
                         val user = ConfigManager.readConfig<UserConfig>(ConfigType.USER, false)?.defaultGithubUser
                             ?: run {
-                                message.error("Default Github User is not set in `${ConfigType.USER.configPath.substring(7)}`!")
+                                message.error(
+                                    "Default Github User is not set in `${
+                                        ConfigType.USER.configPath.substring(
+                                            7
+                                        )
+                                    }`!"
+                                )
                                 return@doesLater
                             }
 
@@ -213,7 +221,8 @@ object IssueCommand : Command("issue") {
         repoName: String,
         issueNum: String,
     ) {
-        val issue = authenticatedRequest<Issue>("token", token, "https://api.github.com/repos/$user/$repoName/issues/$issueNum")
+        val issue =
+            authenticatedRequest<Issue>("token", token, "https://api.github.com/repos/$user/$repoName/issues/$issueNum")
         try {
             if (issue.html_url != null && issue.html_url.contains("issue")) {
                 message.channel.send {
@@ -250,7 +259,8 @@ object IssueCommand : Command("issue") {
             message.channel.send {
                 embed {
                     title = "Error"
-                    description = "Something went wrong when trying to execute this command! Does the user / repo / issue exist?"
+                    description =
+                        "Something went wrong when trying to execute this command! Does the user / repo / issue exist?"
                     field("Stacktrace", "```${e.stackTraceToString()}```", false)
                     e.printStackTrace()
                     color = Colors.ERROR.color
@@ -311,9 +321,9 @@ object IssueCommand : Command("issue") {
 
     override fun getHelpUsage(): String {
         return "Getting information of an issue/pull on github. \n\n" +
-                "Usage: \n" +
-                "`$fullName <user/organization> <repository> <issue>`\n\n" +
-                "Example: \n" +
-                "`$fullName kami-blue bot-kt 10`\n\n"
+            "Usage: \n" +
+            "`$fullName <user/organization> <repository> <issue>`\n\n" +
+            "Example: \n" +
+            "`$fullName kami-blue bot-kt 10`\n\n"
     }
 }
