+++ src/main/bot-kt/org/kamiblue/botkt/command/commands/PurgeCommand.kt
@@ -1,7 +1,8 @@
-package org.kamiblue.botkt.commands
+package org.kamiblue.botkt.command.commands
 
 import net.ayataka.kordis.entity.deleteAll
 import org.kamiblue.botkt.*
+import org.kamiblue.botkt.command.*
 
 object PurgeCommand : Command("purge") {
     init {
@@ -26,7 +27,7 @@ object PurgeCommand : Command("purge") {
 
     override fun getHelpUsage(): String {
         return "Purges a number of messages in a channel based on parameters.\n" +
-                "`$fullName <number>`\n" +
-                "`$fullName <number> <userid>`"
+            "`$fullName <number>`\n" +
+            "`$fullName <number> <userid>`"
     }
 }
