+++ src/main/bot-kt/org/kamiblue/botkt/command/commands/RoleInfoCommand.kt
@@ -1,10 +1,11 @@
-package org.kamiblue.botkt.commands
+package org.kamiblue.botkt.command.commands
 
 import net.ayataka.kordis.entity.message.Message
 import net.ayataka.kordis.entity.server.role.Role
 import org.kamiblue.botkt.*
-import org.kamiblue.botkt.utils.MessageSendUtils.error
+import org.kamiblue.botkt.command.*
 import org.kamiblue.botkt.utils.Colors
+import org.kamiblue.botkt.utils.MessageSendUtils.error
 import org.kamiblue.botkt.utils.pretty
 
 object RoleInfoCommand : Command("roleinfo") {
