+++ src/main/bot-kt/org/kamiblue/botkt/command/commands/StealEmojiCommand.kt
@@ -1,9 +1,6 @@
-package org.kamiblue.botkt.commands
+package org.kamiblue.botkt.command.commands
 
-import org.kamiblue.botkt.Command
-import org.kamiblue.botkt.arg
-import org.kamiblue.botkt.doesLater
-import org.kamiblue.botkt.emoji
+import org.kamiblue.botkt.command.*
 import org.kamiblue.botkt.utils.AnimatableEmoji
 import org.kamiblue.botkt.utils.MessageSendUtils.error
 import org.kamiblue.botkt.utils.MessageSendUtils.normal
