+++ src/main/bot-kt/org/kamiblue/botkt/command/commands/UserInfoCommand.kt
@@ -1,16 +1,12 @@
-package org.kamiblue.botkt.commands
+package org.kamiblue.botkt.command.commands
 
 import net.ayataka.kordis.entity.findByTag
 import net.ayataka.kordis.entity.message.Message
 import net.ayataka.kordis.entity.server.member.Member
-import org.kamiblue.botkt.Command
-import org.kamiblue.botkt.arg
-import org.kamiblue.botkt.doesLater
-import org.kamiblue.botkt.greedyString
+import org.kamiblue.botkt.command.*
 import org.kamiblue.botkt.utils.Colors
 import org.kamiblue.botkt.utils.MessageSendUtils.error
 import org.kamiblue.botkt.utils.ReactionUtils.FakeUser
-import org.kamiblue.botkt.utils.SnowflakeHelper.fromDiscordSnowFlake
 import org.kamiblue.botkt.utils.SnowflakeHelper.prettyFormat
 import org.kamiblue.botkt.utils.SnowflakeHelper.toInstant
 import org.kamiblue.botkt.utils.StringUtils.toHumanReadable
